requirements.txt
discord-webhook
fastapi
pymongo
PyYAML
snscrape
uvicorn
python-dotenv
